+++ lib/cldr.ex
@@ -798,7 +798,7 @@ defmodule Cldr do
 
   ## Options
 
-  * `locale` is any valid locale name returned by `Cldr.known_locale_names/1`.
+  * `:locale` is any valid locale name returned by `Cldr.known_locale_names/1`.
     The default is `Cldr.get_locale/0`
 
   ## Examples
@@ -811,7 +811,7 @@ defmodule Cldr do
 
   """
   @spec quote(String.t(), backend(), Keyword.t()) :: String.t
-  def quote(string, backend, options \\ [])
+  def quote(string, backend \\ default_backend(), options \\ [])
 
   def quote(string, options, []) when is_binary(string) and is_list(options) do
     {backend, options} = Keyword.pop(options, :backend)
@@ -820,8 +820,53 @@ defmodule Cldr do
   end
 
   def quote(string, backend, options) when is_binary(string) and is_list(options) do
-    options = Keyword.merge([locale: default_locale()], options)
     backend.quote(string, options)
+  end
+
+  @doc """
+  Add locale-specific ellipsis to a string.
+
+  ## Arguments
+
+  * `string` is any valid Elixir string
+
+  * `backend` is any module that includes `use Cldr` and therefore
+    is a `Cldr` backend module.  The default is `Cldr.default_backend/0`.
+    Note that `Cldr.default_backend/0` will raise an exception if
+    no `:default_backend` is configured under the `:ex_cldr` key in
+    `config.exs`.
+
+  * `options` is a keyword list of options
+
+  ## Options
+
+  * `:locale` is any valid locale name returned by `Cldr.known_locale_names/1`.
+    The default is `Cldr.get_locale/0`
+
+  * `:location` determines where to place the ellipsis. The options are
+    `:after` (the default for a single string argument), `:between` (the default
+    and only valid location for an argument that is a list of two strings) and `:before`
+
+  ## Examples
+
+      iex> Cldr.ellipsis "And furthermore", MyApp.Cldr
+      "And furthermore…"
+
+      iex> Cldr.ellipsis ["And furthermore", "there is much to be done"], MyApp.Cldr, locale: "ja"
+      "And furthermore…there is much to be done"
+
+  """
+  @spec ellipsis(String.t() | list(String.t()), backend(), Keyword.t()) :: String.t
+  def ellipsis(string, backend \\ default_backend(), options \\ [])
+
+  def ellipsis(string, options, []) when is_list(options) do
+    {backend, options} = Keyword.pop(options, :backend)
+    backend = backend || default_backend()
+    ellipsis(string, backend, options)
+  end
+
+  def ellipsis(string, backend, options) when is_list(options) do
+    backend.ellipsis(string, options)
   end
 
   @doc """