+++ mix/support/consolidate.ex
@@ -104,6 +104,7 @@ defmodule Cldr.Consolidate do
     |> Normalize.LanguageNames.normalize(locale)
     |> Normalize.Calendar.normalize(locale)
     |> Normalize.Delimiter.normalize(locale)
+    |> Normalize.Ellipsis.normalize(locale)
   end
 
   # Remove the top two levels of the map since they add nothing